# src/Http.API/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "dusi.dev",
    "description": "API 文档",
    "version": "v1"
  },
  "paths": {
...
    "/api/EntityModel/filter": {
      "post": {
        "tags": [
          "EntityModel"
        ],
        "operationId": "EntityModel_Filter",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityModelFilterDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityModelFilterDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EntityModelFilterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelItemDtoPageList"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelItemDtoPageList"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelItemDtoPageList"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/EntityModel/{id}": {
      "put": {
        "tags": [
          "EntityModel"
        ],
        "operationId": "EntityModel_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityModelUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityModelUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EntityModelUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModel"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "EntityModel"
        ],
        "operationId": "EntityModel_GetDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModel"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "EntityModel"
        ],
        "operationId": "EntityModel_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/SystemRole/filter": {
      "post": {
        "tags": [
          "SystemRole"
        ],
        "operationId": "SystemRole_Filter",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemRoleFilterDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemRoleFilterDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SystemRoleFilterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SystemRoleItemDtoPageList"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemRoleItemDtoPageList"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemRoleItemDtoPageList"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/SystemRole/{id}": {
      "put": {
        "tags": [
          "SystemRole"
        ],
        "operationId": "SystemRole_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemRoleUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemRoleUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SystemRoleUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SystemRole"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemRole"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemRole"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "SystemRole"
        ],
        "operationId": "SystemRole_GetDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SystemRole"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemRole"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemRole"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "SystemRole"
        ],
        "operationId": "SystemRole_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SystemRole"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemRole"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemRole"
                }
              }
            }
          }
        }
      }
    },
    "/api/SystemUser/filter": {
      "post": {
        "tags": [
          "SystemUser"
        ],
        "operationId": "SystemUser_Filter",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemUserFilterDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemUserFilterDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SystemUserFilterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SystemUserItemDtoPageList"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemUserItemDtoPageList"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemUserItemDtoPageList"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/SystemUser/{id}": {
      "put": {
        "tags": [
          "SystemUser"
        ],
        "operationId": "SystemUser_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemUserUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemUserUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SystemUserUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SystemUser"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemUser"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemUser"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "SystemUser"
        ],
        "operationId": "SystemUser_GetDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SystemUser"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemUser"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemUser"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "SystemUser"
        ],
        "operationId": "SystemUser_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SystemUser"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemUser"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemUser"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AccessModifier": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "description": "访问修饰符",
        "format": "int32",
        "x-enumNames": [
          "Public",
          "Protected",
          "Internal"
        ]
      },
...
      "EntityModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdTime": {
            "type": "string",
            "format": "date-time"
          },
          "updatedTime": {
            "type": "string",
            "format": "date-time"
          },
          "isDeleted": {
            "type": "boolean",
            "description": "软删除"
          },
          "name": {
            "maxLength": 60,
            "type": "string",
            "description": "实体类名",
            "nullable": true
          },
          "comment": {
            "maxLength": 300,
            "type": "string",
            "description": "实体注释内容",
            "nullable": true
          },
          "accessModifier": {
            "$ref": "#/components/schemas/AccessModifier"
          },
          "codeExample": {
            "maxLength": 2000,
            "type": "string",
            "description": "代码示例",
            "nullable": true
          },
          "codeLanguage": {
            "$ref": "#/components/schemas/CodeLanguage"
          },
          "parentEntity": {
            "$ref": "#/components/schemas/EntityModel"
          },
          "childrenEntities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntityModel"
            },
            "description": "直属子类",
            "nullable": true
          },
          "entityMembers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntityMember"
            },
            "description": "包含的属性",
            "nullable": true
          },
          "entityLibrary": {
            "$ref": "#/components/schemas/EntityLibrary"
          }
        },
        "additionalProperties": false,
        "description": "实体模型类"
      },
...
      "EntityModelFilterDto": {
        "type": "object",
        "properties": {
          "pageIndex": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "orderBy": {
            "type": "object",
            "additionalProperties": {
              "type": "boolean",
              "nullable": true
            },
            "description": "排序",
            "nullable": true
          },
          "name": {
            "maxLength": 60,
            "type": "string",
            "description": "实体类名",
            "nullable": true
          },
          "comment": {
            "maxLength": 300,
            "type": "string",
            "description": "实体注释内容",
            "nullable": true
          },
          "accessModifier": {
            "$ref": "#/components/schemas/AccessModifier"
          },
          "parentEntityId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "entityLibraryId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "实体模型类查询筛选"
      },
      "EntityModelItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 60,
            "type": "string",
            "description": "实体类名",
            "nullable": true
          },
          "comment": {
            "maxLength": 300,
            "type": "string",
            "description": "实体注释内容",
            "nullable": true
          },
          "accessModifier": {
            "$ref": "#/components/schemas/AccessModifier"
          },
          "codeLanguage": {
            "$ref": "#/components/schemas/CodeLanguage"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdTime": {
            "type": "string",
            "format": "date-time"
          },
          "updatedTime": {
            "type": "string",
            "format": "date-time"
          },
          "isDeleted": {
            "type": "boolean",
            "description": "软删除"
          }
        },
        "additionalProperties": false,
        "description": "实体模型类列表元素"
      },
      "EntityModelItemDtoPageList": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntityModelItemDto"
            },
            "nullable": true
          },
          "pageIndex": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "EntityModelUpdateDto": {
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 60,
            "type": "string",
            "description": "实体类名",
            "nullable": true
          },
          "comment": {
            "maxLength": 300,
            "type": "string",
            "description": "实体注释内容",
            "nullable": true
          },
          "accessModifier": {
            "$ref": "#/components/schemas/AccessModifier"
          },
          "codeExample": {
            "maxLength": 2000,
            "type": "string",
            "description": "代码示例",
            "nullable": true
          },
          "codeLanguage": {
            "$ref": "#/components/schemas/CodeLanguage"
          },
          "entityLibraryId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false,
        "description": "实体模型类更新时请求结构"
      },
...
      "Sex": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "description": "性别",
        "format": "int32",
        "x-enumNames": [
          "Male",
          "Female",
          "Else"
        ]
      },
      "SystemRole": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdTime": {
            "type": "string",
            "format": "date-time"
          },
          "updatedTime": {
            "type": "string",
            "format": "date-time"
          },
          "isDeleted": {
            "type": "boolean",
            "description": "软删除"
          },
          "name": {
            "maxLength": 30,
            "type": "string",
            "description": "角色显示名称",
            "nullable": true
          },
          "nameValue": {
            "type": "string",
            "description": "角色名，系统标识",
            "nullable": true
          },
          "isSystem": {
            "type": "boolean",
            "description": "是否系统内置,系统内置不可删除"
          },
          "icon": {
            "maxLength": 30,
            "type": "string",
            "description": "图标",
            "nullable": true
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SystemUser"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "角色表"
      },
...
      "SystemRoleFilterDto": {
        "type": "object",
        "properties": {
          "pageIndex": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "orderBy": {
            "type": "object",
            "additionalProperties": {
              "type": "boolean",
              "nullable": true
            },
            "description": "排序",
            "nullable": true
          },
          "name": {
            "maxLength": 30,
            "type": "string",
            "description": "角色显示名称",
            "nullable": true
          },
          "nameValue": {
            "type": "string",
            "description": "角色名，系统标识",
            "nullable": true
          },
          "isSystem": {
            "type": "boolean",
            "description": "是否系统内置,系统内置不可删除",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "角色查询筛选"
      },
...
      "SystemRoleItemDtoPageList": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SystemRoleItemDto"
            },
            "nullable": true
          },
          "pageIndex": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "SystemRoleUpdateDto": {
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 30,
            "type": "string",
            "description": "角色显示名称",
            "nullable": true
          },
          "nameValue": {
            "type": "string",
            "description": "角色名，系统标识",
            "nullable": true
          },
          "isSystem": {
            "type": "boolean",
            "description": "是否系统内置,系统内置不可删除",
            "nullable": true
          },
          "icon": {
            "maxLength": 30,
            "type": "string",
            "description": "图标",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "角色更新时请求结构"
      },
...
      "SystemUserFilterDto": {
        "type": "object",
        "properties": {
          "pageIndex": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "orderBy": {
            "type": "object",
            "additionalProperties": {
              "type": "boolean",
              "nullable": true
            },
            "description": "排序",
            "nullable": true
          },
          "userName": {
            "maxLength": 30,
            "type": "string",
            "description": "用户名",
            "nullable": true
          },
          "emailConfirmed": {
            "type": "boolean",
            "nullable": true
          },
          "phoneNumberConfirmed": {
            "type": "boolean",
            "nullable": true
          },
          "twoFactorEnabled": {
            "type": "boolean",
            "nullable": true
          },
          "lockoutEnabled": {
            "type": "boolean",
            "nullable": true
          },
          "accessFailedCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "retryCount": {
            "type": "integer",
            "description": "密码重试次数",
            "format": "int32",
            "nullable": true
          },
          "sex": {
            "$ref": "#/components/schemas/Sex"
          }
        },
        "additionalProperties": false,
        "description": "系统用户查询筛选"
      },
...
      "SystemUserItemDtoPageList": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SystemUserItemDto"
            },
            "nullable": true
          },
          "pageIndex": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "SystemUserUpdateDto": {
        "type": "object",
        "properties": {
          "userName": {
            "maxLength": 30,
            "type": "string",
            "description": "用户名",
            "nullable": true
          },
          "realName": {
            "maxLength": 30,
            "type": "string",
            "description": "真实姓名",
            "nullable": true
          },
          "email": {
            "maxLength": 100,
            "type": "string",
            "nullable": true
          },
          "emailConfirmed": {
            "type": "boolean",
            "nullable": true
          },
          "phoneNumber": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "phoneNumberConfirmed": {
            "type": "boolean",
            "nullable": true
          },
          "twoFactorEnabled": {
            "type": "boolean",
            "nullable": true
          },
          "lockoutEnd": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lockoutEnabled": {
            "type": "boolean",
            "nullable": true
          },
          "accessFailedCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "lastLoginTime": {
            "type": "string",
            "description": "最后登录时间",
            "format": "date-time",
            "nullable": true
          },
          "retryCount": {
            "type": "integer",
            "description": "密码重试次数",
            "format": "int32",
            "nullable": true
          },
          "avatar": {
            "maxLength": 200,
            "type": "string",
            "description": "头像url",
            "nullable": true
          },
          "sex": {
            "$ref": "#/components/schemas/Sex"
          }
        },
        "additionalProperties": false,
        "description": "系统用户更新时请求结构"
      }
    }
  }
}